+++ src/lib/Backend/Trivia/Trivia.ts
@@ -1,29 +1,7 @@
 import fetch from 'node-fetch';
 import { Question } from './types/Trivia';
 
-const shuffle = <T>(a: T[]): T[] => {
-    for(let i = a.length - 1; i > 0; i--) {
-        const j = Math.floor(Math.random() * (i + 1));
-        [a[i], a[j]] = [a[j], a[i]];
-    }
-    return a;
-}
-
 export const categories: { id: number, name: string }[] = [];
-// push questions to existing array (if exists already)
-// or set the questions to a new array.
-// Proxy is amazing. 
-export const questions = new Proxy({} as { [key: number]: Question[] }, {
-    set: (target, prop: number, value) => {
-        if(prop in target) {
-            target[prop].push(...value);
-        } else {
-            target[prop] = [...value];
-        }
-
-        return true;
-    }
-});
 export let categoryRegex: RegExp = null;
 
 export const Trivia = {
@@ -42,22 +20,25 @@ export const Trivia = {
         return categories;
     },
 
-    fetchQuestions: async (amount: number, category: number, difficulty: 'easy' | 'medium' | 'hard') => {
-        if(category in questions && questions[category].length >= amount) {
-            return shuffle(questions[category]).splice(0, amount);
+    fetchAllQuestions: async () => {
+        const questions: Question[] = [];
+        const token = await (await fetch('https://opentdb.com/api_token.php?command=request')).json();
+
+        if(token.response_code !== 0) {
+            throw token;
         }
 
-        const base = `https://opentdb.com/api.php?amount=50&category=${category}&difficulty=${difficulty}`;
-        const res = await fetch(base);
-
-        if(res.status === 200) {
+        while(true) {
+            const res = await fetch(`https://opentdb.com/api.php?amount=50&token=${token.token}`);
             const json = await res.json();
+
             if(json.response_code === 0) {
-                questions[category] = json.results;
-                return shuffle(questions[category]).splice(0, amount);
+                questions.push(...json.results);
             } else {
-                return null;
+                break;
             }
         }
+        
+        return questions;
     }
 }